# app/backend/cmd/api/docs/swagger.yaml
definitions:
  common.ErrorResponse:
    properties:
      error:
        description: 'example: user not found'
        type: string
    type: object
...
  handlers.CreateDefectRequest:
    properties:
      building_id:
        description: 'example: 1'
        type: integer
      deadline:
        description: |-
          example: 2025-12-31 23:59:59
          deadline in layout "2006-01-02 15:04:05"
        type: string
      description:
        description: 'example: Описание дефекта...'
        type: string
      priority:
        description: 'example: high'
        type: string
      responsible_person_id:
        description: 'example: 2'
        type: integer
      status:
        description: 'example: new'
        type: string
      title:
        description: 'example: Трещина в стене'
        type: string
    type: object
...
  handlers.DefectResponse:
    properties:
      building:
        $ref: '#/definitions/handlers.SimpleBuilding'
      building_id:
        type: integer
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/handlers.SimpleUser'
      created_by_person_id:
        type: integer
      deadline:
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        type: string
      responsible:
        $ref: '#/definitions/handlers.SimpleUser'
      responsible_person_id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      updated_by_person_id:
        type: integer
    type: object
...
  handlers.SimpleBuilding:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      stage:
        type: string
    type: object
  handlers.SimpleUser:
    properties:
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: this is documentation for buildefect API
  title: buildefect api
  version: "1.0"
paths:
...
  /api/defects:
    get:
      consumes:
      - application/json
      description: Retrieve defects with optional filters and pagination
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by building id
        in: query
        name: building_id
        type: integer
      - description: Filter by responsible user id
        in: query
        name: responsible_id
        type: integer
      - description: Limit number of results (default 100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.DefectResponse'
            type: array
        "400":
          description: invalid query param
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List defects
      tags:
      - defects
    post:
      consumes:
      - application/json
      description: Create a defect. Requires authentication.
      parameters:
      - description: Defect payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDefectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.DefectResponse'
        "400":
          description: invalid request body or missing fields
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "401":
          description: unauthenticated
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create defect
      tags:
      - defects
  /api/defects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete defect by id
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Successfully deleted defect with id {id}
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete defect
      tags:
      - defects
    get:
      consumes:
      - application/json
      description: Retrieve defect by numeric id
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DefectResponse'
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: defect not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get defect by id
      tags:
      - defects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"